implementing unit test #1
