+++ package.json
@@ -12,6 +12,7 @@
     "format": "prettier --write src/",
     "test:e2e": "playwright test",
     "test:e2e:ui": "playwright test --ui",
+    "test" : "vitest",
     "type-check": "vue-tsc --noEmit -p tsconfig.app.json --composite false"
   },
   "dependencies": {
